created func that generates token which will be sent to client
