+++ app.js
@@ -11,7 +11,7 @@ const swaggerUI = require('swagger-ui-express')
 
 const swaggerOptions = require('~/swagger-settings')
 const example = require('~/routes/example')
-const authRoutes = require('~/routes/auth')
+const auth = require('~/routes/auth')
 
 const app = express()
 
@@ -28,7 +28,7 @@ const swaggerSettings = swaggerJsDoc(swaggerOptions)
 app.use('/api-docs',swaggerUI.serve,swaggerUI.setup(swaggerSettings))
 
 app.use('/example', example)
-app.use(authRoutes) // localhost:8080/...
+app.use('/auth', auth) // localhost:8080/auth/...
 
 mongoose
   .connect(`mongodb+srv://${process.env.MONGODB_USERNAME}:${process.env.MONGODB_PASSWORD}@teachma.693y8.mongodb.net/test`)
